[Gold IV] Title: 기차 여행, Time: 412 ms, Memory: 58288 KB, Score: 100 point -BaekjoonHub

+++ 백준/Gold/10713. 기차 여행/기차 여행.java
@@ -49,10 +49,10 @@ public class Main {
 			C[i] = Integer.parseInt(st.nextToken());
 		}
 		
-		int answer = 0;
-		int prevCnt = 0;
+		long answer = 0;
+		long prevCnt = 0;
 		for(int i = min; i < max; i++) {
-			int cnt = prevCnt+arr[i];
+			long cnt = prevCnt+arr[i];
 			prevCnt = cnt;
 			answer += Math.min(A[i]*cnt, B[i]*cnt+C[i]);
 		}
